Renames ethereum transactions to evm
