+++ frontend/app/src/auto-imports.d.ts
@@ -324,6 +324,7 @@ declare global {
   const useTooltip: typeof import('./composables/graphs')['useTooltip']
   const useTradeFilters: typeof import('./composables/filters/trades')['useTradeFilters']
   const useTransactionFilter: typeof import('./composables/filters/transactions')['useTransactionFilter']
+  const useTransactionQueryStatus: typeof import('./composables/txs')['useTransactionQueryStatus']
   const useTransition: typeof import('@vueuse/core')['useTransition']
   const useTrunc: typeof import('@vueuse/math')['useTrunc']
   const useUrlSearchParams: typeof import('@vueuse/core')['useUrlSearchParams']
@@ -683,6 +684,7 @@ declare module 'vue' {
     readonly useTooltip: UnwrapRef<typeof import('./composables/graphs')['useTooltip']>
     readonly useTradeFilters: UnwrapRef<typeof import('./composables/filters/trades')['useTradeFilters']>
     readonly useTransactionFilter: UnwrapRef<typeof import('./composables/filters/transactions')['useTransactionFilter']>
+    readonly useTransactionQueryStatus: UnwrapRef<typeof import('./composables/txs')['useTransactionQueryStatus']>
     readonly useTransition: UnwrapRef<typeof import('@vueuse/core')['useTransition']>
     readonly useTrunc: UnwrapRef<typeof import('@vueuse/math')['useTrunc']>
     readonly useUrlSearchParams: UnwrapRef<typeof import('@vueuse/core')['useUrlSearchParams']>
